+++ src/benchmarklib/tpcc/procedures/tpcc_payment.cpp
@@ -31,7 +31,7 @@ TPCCPayment::TPCCPayment(const int num_warehouses, BenchmarkSQLExecutor& sql_exe
   if (select_customer_by_name) {
     customer = pmr_string{_tpcc_random_generator.last_name(_tpcc_random_generator.nurand(255, 0, 999))};
   } else {
-    customer = _tpcc_random_generator.nurand(1023, 1, 3000);
+    customer = static_cast<int32_t>(_tpcc_random_generator.nurand(1023, 1, 3000));
   }
 
   // Generate payment information
